Adding slotted messages in TreeRoot.vue
Rethinking my previous PR: it is better to have wonderful component slot to display message on empty filtering result. We can get rid of emptyText prop and custom div with custom class.
And along with this: it is good to have slot for empty dataset message. Because when we fetch data from external source - we never know if there is actually any structure. So I added another slot on matches.length == 0